app/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
choicelib==0.1.5
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
frozenlist==1.8.0
greenlet==3.3.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.9.0
msgspec==0.19.0
multidict==6.7.1
numpy==2.4.2
packaging==26.0
pandas==2.2.2
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.32
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
vbml==1.1.post1
vkbottle==4.6.2
vkbottle-types==5.199.99.12
yarl==1.22.0
